fix(系统设置): 调整位置和去掉localStorage
