+++ frontend/src/App.vue
@@ -14,10 +14,12 @@
   import { saveBaseInfo, getBaseInfo } from '@/api/modules/setting/config';
   import { getLocalStorage, setLocalStorage } from '@/utils/local-storage';
   import useAppStore from '@/store/modules/app';
+  import useLicenseStore from '@/store/modules/setting/license';
   import { watchStyle, watchTheme, setFavicon } from '@/utils/theme';
   import { GetPlatformIconUrl } from '@/api/requrls/setting/config';
 
   const appStore = useAppStore();
+  const licenseStore = useLicenseStore();
 
   const { currentLocale } = useLocale();
   const locale = computed(() => {
@@ -41,6 +43,7 @@
     try {
       appStore.initSystemversion(); // 初始化系统版本
       appStore.initPageConfig(); // 初始化页面配置
+      licenseStore.getValidateLicense(); // 初始化校验license
       // 项目初始化时需要获取基础设置信息，看当前站点 url是否为系统内置默认地址，如果是需要替换为当前项目部署的 url 地址
       const isInitUrl = getLocalStorage('isInitUrl'); // 是否已经初始化过 url
       if (isInitUrl === 'true') return;
